+++ src/components/Buttons/Button/index.tsx
@@ -1,5 +1,6 @@
+import classNames from "classnames";
+import styles from "../Buttons.module.css";
 import { ButtonProps } from "./@types";
-import styles from "./Button.module.css";
 import { Loading } from "@components/Loading";
 
 export function Button(props: ButtonProps) {
@@ -13,16 +14,16 @@ export function Button(props: ButtonProps) {
     onClick,
   } = props;
 
-  const classes = `
-    button 
-    ${loading ? styles.loading : ""} 
-    ${disabled ? styles.disabled : ""} 
-    ${className || ""}
-  `;
+  const classes = classNames(
+    styles.button,
+    className,
+    loading && styles.loading,
+    disabled && styles.disabled,
+  );
 
   return (
     <button onClick={onClick} type={type || "button"} className={classes}>
-      {loading ? <Loading className={loadingClassName || "w-6"} /> : children}
+      {loading ? <Loading className={loadingClassName || "w-5"} /> : children}
     </button>
   );
 }