+++ src/components/Input/@types/index.ts
@@ -1,11 +1,14 @@
+import { InputHTMLAttributes } from "react";
 import { UseFormRegisterReturn } from "react-hook-form";
 
 export type InputTypes = "text" | "password";
 
-export type InputProps = {
+export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   type: InputTypes;
   name: string;
   label: string;
   placeholder: string;
   register?: UseFormRegisterReturn;
-};
+  messageType?: "error" | "success";
+  message?: string;
+}