+++ src/components/Input/index.tsx
@@ -2,10 +2,23 @@ import { InputProps } from "./@types";
 import styles from "./Input.module.css";
 
 export function Input(props: InputProps) {
-  const { type, name, placeholder, label, register } = props;
+  const {
+    type,
+    name,
+    placeholder,
+    label,
+    register,
+    message,
+    messageType,
+    ...others
+  } = props;
+
+  // TODO: fix styles type
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const wrapperClasses = `${styles.wrapper} ${styles[messageType as any]}`;
 
   return (
-    <div>
+    <div className={wrapperClasses}>
       {label && (
         <label className={styles.label} htmlFor={name}>
           {label}
@@ -18,7 +31,9 @@ export function Input(props: InputProps) {
         placeholder={placeholder}
         className={styles.input}
         {...register}
+        {...others}
       />
+      {<small className={styles.message}>{message}</small>}
     </div>
   );
 }