+++ src/components/Toggle/@types/index.ts
@@ -1,0 +1,6 @@
+export type ToggleProps = {
+  label?: string;
+  className?: string;
+  checked?: boolean;
+  onChange: (isPrivate: boolean) => void;
+};