+++ src/features/AddPost/index.tsx
@@ -3,7 +3,7 @@ import styles from "./AddPost.module.css";
 import { Counter } from "./components/Counter";
 import { Header } from "./components/Header";
 import { useAddPost } from "./hooks/useAddPost";
-import { Button } from "@components/Button";
+import { Button } from "@components/Buttons/Button";
 import { CONTENT_MAX_LENGTH } from "@config";
 
 export function AddPost() {
@@ -11,7 +11,7 @@ export function AddPost() {
   const {
     register,
     handleSubmit,
-    formState: { errors },
+    formState: { isValid },
     watch,
   } = formMethods;
   const { t } = useTranslation("addPost");
@@ -44,7 +44,7 @@ export function AddPost() {
             <Button
               className={styles.submit}
               loading={loading}
-              disabled={!!errors.content?.message}
+              disabled={!isValid}
               type="submit"
             >
               {t("buttons.addPost")}