+++ src/features/FollowRequests/index.tsx
@@ -1,11 +1,12 @@
 import styles from "./FollowRequests.module.css";
 import { useFollowRequests } from "./hooks/useFollowRequests";
+import { Card } from "@components/Card";
 import { UserInfo } from "@features/UserInfo";
 
 export function FollowRequests() {
   const { followRequests } = useFollowRequests();
   return (
-    <div className={styles.card}>
+    <Card className={styles.card}>
       {followRequests?.map(followRequest => (
         <UserInfo
           key={followRequest.id}
@@ -14,6 +15,6 @@ export function FollowRequests() {
           showDelete
         />
       ))}
-    </div>
+    </Card>
   );
 }