+++ src/features/Profile/components/Header.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from "react-i18next";
 import styles from "../Profile.module.css";
 import { useProfile } from "../hooks";
 import { Avatar } from "@components/Avatar";
-import { LinkButton } from "@components/LinkButton";
+import { LinkButton } from "@components/Buttons/LinkButton";
 import { useAccountContext } from "@context/AccountContext";
 import { PageUrls } from "@enums/pages";
 import { ConnectionStatus, Count } from ".";
@@ -11,7 +11,8 @@ import type { User } from "@t/user";
 export function Header({
   id,
   username,
-  fullname,
+  fullName,
+  bio,
   connectionStatus,
   isPrivate,
 }: User) {
@@ -27,9 +28,13 @@ export function Header({
       <Avatar className={styles.avatar} />
       <div className="flex items-start justify-between w-full">
         <div>
-          <div className={styles.fullname}>{fullname}</div>
+          <div className={styles.fullName}>{fullName}</div>
           <div className={styles.username}>@{username}</div>
         </div>
+      </div>
+
+      <div className="w-full">
+        <p className={styles.bio}>{bio}</p>
       </div>
       <div className="w-full flex items-end justify-between">
         <div className="flex w-full text-tiny space-x-3">