+++ src/features/Settings/hooks/index.ts
@@ -1,0 +1,5 @@
+import { useEditUser } from "./useEditUser";
+import { useEditUserValidations } from "./useEditUserValidations";
+import { useSettings } from "./useSettings";
+
+export { useSettings, useEditUser, useEditUserValidations };