+++ src/graphql/@types/user.ts
@@ -4,6 +4,18 @@ export type GetUserInfoQuery = {
   getUserInfo: User;
 };
 
+export type CheckUsernameQuery = {
+  checkUsername: User;
+};
+
+export type CheckEmailQuery = {
+  checkEmail: User;
+};
+
 export type GetUserByUsernameQuery = {
   getUserByUsername: User;
+};
+
+export type EditUserMutation = {
+  editUser: User;
 };