+++ src/graphql/user.ts
@@ -6,11 +6,29 @@ export const GET_USER_INFO = gql`
       id
       username
       email
-      fullname
+      fullName
+      bio
+      isPrivate
       followersCount
       followingCount
       followRequestsCount
       postsCount
+    }
+  }
+`;
+
+export const CHECK_USERNAME = gql`
+  query checkUsername($username: String!) {
+    checkUsername(username: $username) {
+      id
+    }
+  }
+`;
+
+export const CHECK_EMAIL = gql`
+  query checkEmail($email: String!) {
+    checkEmail(email: $email) {
+      id
     }
   }
 `;
@@ -21,11 +39,43 @@ export const GET_USER_BY_USERNAME = gql`
       id
       username
       email
-      fullname
+      fullName
+      bio
       isPrivate
       connectionStatus
       followersCount
       followingCount
+      postsCount
+    }
+  }
+`;
+
+export const EDIT_USER = gql`
+  mutation editUser(
+    $username: String
+    $fullName: String
+    $email: String
+    $bio: String
+    $isPrivate: Boolean
+  ) {
+    editUser(
+      input: {
+        username: $username
+        fullName: $fullName
+        email: $email
+        bio: $bio
+        isPrivate: $isPrivate
+      }
+    ) {
+      id
+      username
+      email
+      fullName
+      bio
+      isPrivate
+      followersCount
+      followingCount
+      followRequestsCount
       postsCount
     }
   }